Move variables to cartEvents.js to avoid circular dependency
