+++ src/common/lib/cartEvents.js
@@ -1,0 +1,2 @@
+export const giftAddedEvent = 'giftAddedToCart'
+export const cartUpdatedEvent = 'cartUpdatedEvent'
